Add formula to install and run Atlassian Confluence
Confluence is similar to JIRA in most respects, and this formula is
mostly a copy of that for JIRA. The nuances specific to Confluence
are the `confluence-init.properties`.

`service confluence status` will not likely tell you the status, but
you can use this to start/stop the service.

This formula supports the `ssl` and `proxy` pillar keys, just as the
formula for JIRA.

This is the first major step towards completing #5.